+++ src/main/java/com/ydo4ki/bullsandcows/Main.java
@@ -1,5 +1,6 @@
 package com.ydo4ki.bullsandcows;
 
+import java.util.InputMismatchException;
 import java.util.Scanner;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -11,7 +12,15 @@ public class Main {
 		Scanner scanner = new Scanner(System.in);
 		
 		System.out.println("Welcome! Enter amount of digits to guess: ");
-		int digits = scanner.nextInt();
+		int digits;
+		while(true) try {
+			String line = scanner.nextLine();
+			digits = Integer.parseInt(line);
+			break;
+		} catch (NumberFormatException e) {
+			System.out.println("Invalid input :(");
+			continue;
+		}
 		String number = generateNumber(digits);
 		if (DEBUG) System.out.println(number);