+++ changelog.md
@@ -1,4 +1,5 @@
 ## 1.1
+* Fixed a bug due to which program crashes on invalid number input
 * Fixed a grammatical error with bulls and cows plural output
 * Removed debug output that gave out the guessed number so game makes sense now